% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGBoostSub_bin.R
\name{eval_metric_bin}
\alias{eval_metric_bin}
\title{Evaluation Metrics for XGBoostSub_bin Model}
\usage{
eval_metric_bin(model, X_feature, y_label, pi, trt, Loss_type = "A_learning")
}
\arguments{
\item{model}{The trained XGBoostSub_bin model object.}

\item{X_feature}{The input features matrix.}

\item{y_label}{The input y matrix.}

\item{pi}{The propensity scores vector, which should range from 0 to 1, representing the probability of assignment to treatment.}

\item{trt}{The treatment indicator vector. Should take values of 1 or -1, where 1 represents the treatment group and -1 represents the control group.}

\item{Loss_type}{Type of loss function to use: "A_learning" or "Weight_learning".}
}
\value{
Evaluation result of the XGBoostSub_bin model.
}
\description{
Function for evaluating XGBoostSub_bin model performance.
}
\details{
eval_metric: Function for Evaluating XGBoostSub_bin Model Performance

This function evaluates the performance of an XGBoostSub_bin model using a A-learning or weight-learning function.
}
