% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGBoostSub_sur.R
\name{eval_metric_sur}
\alias{eval_metric_sur}
\title{Evaluation Metrics for XGBoostSub_sur Model}
\usage{
eval_metric_sur(
  model,
  X_feature,
  y_label,
  pi,
  trt,
  censor,
  Loss_type = "A_learning"
)
}
\arguments{
\item{model}{The trained XGBoostSub_sur model object.}

\item{X_feature}{The input features matrix.}

\item{y_label}{The input y matrix.}

\item{pi}{The propensity scores vector, which should range from 0 to 1, representing the probability of assignment to treatment.}

\item{trt}{The treatment indicator vector. Should take values of 1 or -1, where 1 represents the treatment group and -1 represents the control group.}

\item{censor}{The censor status vector. Should take values of 1 or 0, where 1 represents censoring and 0 represents an observed event.}

\item{Loss_type}{Type of loss function to use: "A_learning" or "Weight_learning".}
}
\value{
Evaluation result of the XGBoostSub_sur model.
}
\description{
Function for evaluating XGBoostSub_sur model performance.
}
\details{
eval_metric: Function for Evaluating XGBoostSub_con Model Performance

This function evaluates the performance of an XGBoostSub_con model using a A-learning or weight-learning function.
}
