% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{isout}
\alias{isout}
\title{Outlier Detection Based on Boxplot}
\usage{
isout(x)
}
\arguments{
\item{x}{A numeric vector of values for outlier detection.}
}
\value{
A logical vector indicating which elements of the input vector are outliers (TRUE if an outlier, FALSE otherwise).
}
\description{
This function identifies outliers in a numeric vector based on the interquartile range (IQR) method used in boxplots.
}
\keyword{internal}
