% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\docType{data}
\name{tutorial_data}
\alias{tutorial_data}
\title{Tutorial Data}
\format{
A data frame with the following columns:
\describe{
  \item{x1}{Numeric. A biomarker variable.}
  \item{x2}{Numeric. A biomarker variable.}
  \item{x3}{Numeric. A biomarker variable.}
  \item{x4}{Numeric. A biomarker variable.}
  \item{x5}{Numeric. A biomarker variable.}
  \item{x6}{Numeric. A biomarker variable.}
  \item{x7}{Numeric. A biomarker variable.}
  \item{x8}{Numeric. A biomarker variable.}
  \item{x9}{Numeric. A biomarker variable.}
  \item{x10}{Numeric. A biomarker variable.}
  \item{y.con}{Numeric. A continuous outcome variable.}
  \item{y.bin}{Binary. A binary outcome variable, where 0 represents one class and 1 represents another class.}
  \item{y.time}{Numeric. The time in months, used for survival analysis.}
  \item{y.event}{Binary. Event indicator variable, where 0 indicates censoring and 1 indicates the event of interest occurred.}
  \item{subgroup_label}{Binary. Ground truth of subgroup label. In real-world scenarios, this information is typically unavailable.}
  \item{treatment}{Binary. Treatment indicator variable, where 0 represents control and 1 represents treatment.}
  \item{treatment_categorical}{Factor. A categorical version of the treatment variable, with levels "Placebo" and "Treatment".}
  \item{risk_category}{Factor.}
}
}
\usage{
data(tutorial_data)
}
\description{
A dataset containing sample data for demonstrating the functionalities of the BioPred package.
}
\details{
This dataset is used to illustrate various functions within the BioPred package, including predictive modeling and subgroup analysis. The columns represent different types of data typically encountered in clinical studies.
}
\examples{
data(tutorial_data)
head(tutorial_data)
}
\keyword{datasets}
