% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{cdf_plot}
\alias{cdf_plot}
\title{CDF  Plot for a biomarker}
\usage{
cdf_plot(xvar, data, y.int = 5, xlim = NULL, xvar.display = xvar, group = NULL)
}
\arguments{
\item{xvar}{The biomarker name.}

\item{data}{The dataset.}

\item{y.int}{Increasement interval on the y.}

\item{xlim}{cdf plot range for xvar, when NULL, c(min(x), max(x)) will be used.}

\item{xvar.display}{Display name of the biomarker.}

\item{group}{A separate CDF line will be plotted for each group.}
}
\value{
A ggplot object representing the CDF inverse plot.
}
\description{
Cumulative Distribution Function (CDF) plot for a biomarker.
}
