% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{predictive_biomarker_imp}
\alias{predictive_biomarker_imp}
\title{Plot Predictive Biomarker Importance based on XGBoost-based Subgroup Model}
\usage{
predictive_biomarker_imp(model)
}
\arguments{
\item{model}{The trained XGBoost-based model.}
}
\value{
A barplot showing the biomarker importance.
}
\description{
This function calculates and plots the importance of biomarkers in a trained XGBoostSub_con, XGBoostSub_bin or XGBoostSub_sur model.
}
